restored default team exclusion

+++ src/Controller/Api/Team/TeamController.php
@@ -123,6 +123,10 @@ class TeamController extends AbstractController
      */
     public function team(Team $team, ViewFactoryContext $viewFactoryContext)
     {
+        if (Team::DEFAULT_GROUP_ALIAS === $team->getAlias()) {
+            throw new \LogicException('Illegal team');
+        }
+
         $this->denyAccessUnlessGranted(UserTeamVoter::USER_TEAM_VIEW, $team);
         $teamView = $viewFactoryContext->view($team);
 
@@ -160,9 +164,9 @@ class TeamController extends AbstractController
         /** @var User $user */
         $user = $this->getUser();
         if ($user->hasRole(User::ROLE_ADMIN) || $user->hasRole(User::ROLE_SUPER_ADMIN)) {
-            $teams = $teamRepository->findAll();
+            $teams = $teamRepository->findAllExceptDefault();
         } else {
-            $teams = $teamRepository->findByUser($this->getUser());
+            $teams = $teamRepository->findByUserExceptDefault($this->getUser());
         }
 
         $teamView = $viewFactoryContext->viewList($teams);
@@ -203,6 +207,10 @@ class TeamController extends AbstractController
      */
     public function update(Team $team, Request $request, ViewFactoryContext $viewFactoryContext, EntityManagerInterface $entityManager)
     {
+        if (Team::DEFAULT_GROUP_ALIAS === $team->getAlias()) {
+            throw new \LogicException('Illegal team');
+        }
+
         $this->denyAccessUnlessGranted(UserTeamVoter::USER_TEAM_EDIT, $team);
 
         $form = $this->createForm(EditTeamType::class, $team);
@@ -240,6 +248,10 @@ class TeamController extends AbstractController
      */
     public function delete(Team $team, EntityManagerInterface $entityManager)
     {
+        if (Team::DEFAULT_GROUP_ALIAS === $team->getAlias()) {
+            throw new \LogicException('Illegal team');
+        }
+
         $this->denyAccessUnlessGranted(TeamVoter::TEAM_CREATE, $this->getUser());
         $entityManager->remove($team);
         $entityManager->flush();
@@ -274,6 +286,10 @@ class TeamController extends AbstractController
      */
     public function members(Request $request, Team $team, UserTeamRepository $userTeamRepository)
     {
+        if (Team::DEFAULT_GROUP_ALIAS === $team->getAlias()) {
+            throw new \LogicException('Illegal team');
+        }
+
         $this->denyAccessUnlessGranted(UserTeamVoter::USER_TEAM_VIEW, $team);
         $ids = $request->query->get('ids', []);
         $usersTeams = $userTeamRepository->findMembers($team, $ids);
@@ -311,6 +327,10 @@ class TeamController extends AbstractController
      */
     public function addMember(Request $request, Team $team, User $user, EntityManagerInterface $entityManager)
     {
+        if (Team::DEFAULT_GROUP_ALIAS === $team->getAlias()) {
+            throw new \LogicException('Illegal team');
+        }
+
         $this->denyAccessUnlessGranted(UserTeamVoter::USER_TEAM_EDIT, $team);
         $userTeam = new UserTeam();
         $form = $this->createForm(AddMemberType::class, $userTeam);
@@ -347,6 +367,10 @@ class TeamController extends AbstractController
      */
     public function removeMember(Team $team, User $user, UserTeamRepository $userTeamRepository): JsonResponse
     {
+        if (Team::DEFAULT_GROUP_ALIAS === $team->getAlias()) {
+            throw new \LogicException('Illegal team');
+        }
+
         $userTeam = $userTeamRepository->findOneByUserAndTeam($user, $team);
         if (!$userTeam instanceof UserTeam) {
             throw new NotFoundHttpException('User Team not found');
@@ -391,6 +415,10 @@ class TeamController extends AbstractController
      */
     public function editMember(Request $request, Team $team, User $user, UserTeamRepository $userTeamRepository)
     {
+        if (Team::DEFAULT_GROUP_ALIAS === $team->getAlias()) {
+            throw new \LogicException('Illegal team');
+        }
+
         $this->denyAccessUnlessGranted(UserTeamVoter::USER_TEAM_EDIT, $team);
 
         $userTeam = $userTeamRepository->findOneByUserAndTeam($user, $team);
@@ -422,6 +450,10 @@ class TeamController extends AbstractController
      */
     public function lists(Team $team, ViewFactoryContext $viewFactoryContext)
     {
+        if (Team::DEFAULT_GROUP_ALIAS === $team->getAlias()) {
+            throw new \LogicException('Illegal team');
+        }
+
         $lists = $viewFactoryContext->viewList($team->getLists()->getChildLists()->toArray());
         array_push($lists, $viewFactoryContext->view($team->getInbox()));
         array_push($lists, $viewFactoryContext->view($team->getTrash()));
